app/requirements.txt
Flask
python-docx
PyYAML
